# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Interactive API for testing Neuron keylib functionality",
        "title": "Keylib API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/signing/sign-digest": {
            "post": {
                "description": "Sign a 32-byte digest with a private key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signing"
                ],
                "summary": "Sign a pre-hashed digest",
                "parameters": [
                    {
                        "description": "Sign digest request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.SignDigestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.SignatureResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/signing/sign-message": {
            "post": {
                "description": "Sign a message with a private key (message is hashed with Keccak256)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signing"
                ],
                "summary": "Sign a message",
                "parameters": [
                    {
                        "description": "Sign message request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.SignMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.SignatureResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/signing/verify-digest": {
            "post": {
                "description": "Verify a signature over a 32-byte digest using a public key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signing"
                ],
                "summary": "Verify a digest signature",
                "parameters": [
                    {
                        "description": "Verify digest request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.VerifyDigestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.VerifyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/signing/verify-message": {
            "post": {
                "description": "Verify a signature over a message using a public key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signing"
                ],
                "summary": "Verify a message signature",
                "parameters": [
                    {
                        "description": "Verify message request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.VerifyMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.VerifyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid hex format"
                }
            }
        },
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.SignDigestRequest": {
            "type": "object",
            "required": [
                "digestHex",
                "privateKeyHex"
            ],
            "properties": {
                "digestHex": {
                    "type": "string",
                    "example": "0x1234567890abcdef...32bytes..."
                },
                "privateKeyHex": {
                    "type": "string",
                    "example": "0x1234567890abcdef..."
                }
            }
        },
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.SignMessageRequest": {
            "type": "object",
            "required": [
                "message",
                "privateKeyHex"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Hello, World!"
                },
                "privateKeyHex": {
                    "type": "string",
                    "example": "0x1234567890abcdef..."
                }
            }
        },
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.SignatureResponse": {
            "type": "object",
            "properties": {
                "r": {
                    "type": "string",
                    "example": "0x..."
                },
                "s": {
                    "type": "string",
                    "example": "0x..."
                },
                "signatureEthHex": {
                    "type": "string",
                    "example": "0x...65bytes..."
                },
                "signatureHex": {
                    "type": "string",
                    "example": "0x...65bytes..."
                },
                "v": {
                    "type": "integer",
                    "example": 0
                },
                "vEthereum": {
                    "type": "integer",
                    "example": 27
                }
            }
        },
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.VerifyDigestRequest": {
            "type": "object",
            "required": [
                "digestHex",
                "publicKeyHex",
                "signatureHex"
            ],
            "properties": {
                "digestHex": {
                    "type": "string",
                    "example": "0x1234567890abcdef...32bytes..."
                },
                "publicKeyHex": {
                    "type": "string",
                    "example": "0x02759b048e7ccf6ba68f9658105a4a139b5f9f5dfd451857c600cc28f33a1a99ae"
                },
                "signatureHex": {
                    "type": "string",
                    "example": "0x...65bytes..."
                }
            }
        },
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.VerifyMessageRequest": {
            "type": "object",
            "required": [
                "message",
                "publicKeyHex",
                "signatureHex"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Hello, World!"
                },
                "publicKeyHex": {
                    "type": "string",
                    "example": "0x02759b048e7ccf6ba68f9658105a4a139b5f9f5dfd451857c600cc28f33a1a99ae"
                },
                "signatureHex": {
                    "type": "string",
                    "example": "0x...65bytes..."
                }
            }
        },
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.VerifyResponse": {
            "type": "object",
            "properties": {
                "valid": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    }
}